[repo] Flatten ast folder (part 4)
