+++ samlang-core/src/ast/asm-program.ts
@@ -1,6 +1,6 @@
 import { AssemblyInstruction, assemblyInstructionToString } from './asm-instructions';
 import { ENCODED_COMPILED_PROGRAM_MAIN } from './common-names';
-import { GlobalVariable } from './common/structs';
+import type { GlobalVariable } from './common-nodes';
 
 export type AssemblyProgram = {
   readonly globalVariables: readonly GlobalVariable[];
