--- samlang-core/src/ast/common/structs.ts
@@ -1,40 +1,0 @@
-import type { ReadonlyHashMap } from '../../util/collections';
-import type { Type, Range, ModuleReference } from '../common-nodes';
-
-/** A common interface for all AST nodes. */
-export interface Node {
-  /** The range of the entire node. */
-  readonly range: Range;
-}
-
-export interface ModuleMembersImport extends Node {
-  readonly importedMembers: readonly (readonly [string, Range])[];
-  readonly importedModule: ModuleReference;
-  readonly importedModuleRange: Range;
-}
-
-export interface GlobalVariable {
-  readonly name: string;
-  readonly content: string;
-}
-
-export interface AnnotatedParameter {
-  readonly name: string;
-  readonly nameRange: Range;
-  readonly type: Type;
-  readonly typeRange: Range;
-}
-
-export interface FieldType {
-  readonly type: Type;
-  readonly isPublic: boolean;
-}
-
-export interface TypeDefinition extends Node {
-  readonly type: 'object' | 'variant';
-  /** A list of fields. Used for ordering during codegen. */
-  readonly names: readonly string[];
-  readonly mappings: Readonly<Record<string, FieldType | undefined>>;
-}
-
-export type Sources<M> = ReadonlyHashMap<ModuleReference, M>;
