+++ samlang-core/src/ast/samlang-expressions.ts
@@ -11,9 +11,9 @@ import {
   TupleType,
   FunctionType,
   Range,
+  Node,
 } from './common-nodes';
 import type { BinaryOperator } from './common-operators';
-import type { Node } from './common/structs';
 import type { Pattern } from './samlang-pattern';
 
 interface BaseExpression extends Node {
