+++ samlang-core/src/ast/samlang-toplevel.ts
@@ -1,11 +1,12 @@
-import type { FunctionType, Range } from './common-nodes';
-import type {
-  AnnotatedParameter,
-  ModuleMembersImport,
-  TypeDefinition,
-  Node,
-} from './common/structs';
+import type { Type, FunctionType, Range, ModuleReference, Node } from './common-nodes';
 import { SamlangExpression } from './samlang-expressions';
+
+type AnnotatedVariable = {
+  readonly name: string;
+  readonly nameRange: Range;
+  readonly type: Type;
+  readonly typeRange: Range;
+};
 
 export interface ClassMemberDeclaration extends Node {
   readonly isPublic: boolean;
@@ -14,11 +15,23 @@ export interface ClassMemberDeclaration extends Node {
   readonly name: string;
   readonly typeParameters: readonly string[];
   readonly type: FunctionType;
-  readonly parameters: readonly AnnotatedParameter[];
+  readonly parameters: readonly AnnotatedVariable[];
 }
 
 export interface ClassMemberDefinition extends ClassMemberDeclaration {
   readonly body: SamlangExpression;
+}
+
+export interface FieldType {
+  readonly type: Type;
+  readonly isPublic: boolean;
+}
+
+export interface TypeDefinition extends Node {
+  readonly type: 'object' | 'variant';
+  /** A list of fields. Used for ordering during codegen. */
+  readonly names: readonly string[];
+  readonly mappings: Readonly<Record<string, FieldType | undefined>>;
 }
 
 export interface ClassInterface<M extends ClassMemberDeclaration = ClassMemberDeclaration>
@@ -33,6 +46,12 @@ export interface ClassInterface<M extends ClassMemberDeclaration = ClassMemberDe
 
 export interface ClassDefinition extends ClassInterface<ClassMemberDefinition> {
   readonly typeDefinition: TypeDefinition;
+}
+
+export interface ModuleMembersImport extends Node {
+  readonly importedMembers: readonly (readonly [string, Range])[];
+  readonly importedModule: ModuleReference;
+  readonly importedModuleRange: Range;
 }
 
 export interface SamlangModule {
